libwasi_littlefs: separate mount/umount logic
